Extract `PuoriborState` tests to a new file
